Fix UMRs indicated by valgrind (x86_64)
See issue #453 - The valgrind tool can be used to detect errors such as
uninitialized memory reads, etc, at runtime. This is similar to what ASAN
does, but does not require compile-time intervention (at the expense of
a far greater runtime performance impact)

The `write_validate` function, in conjunction with one of the
`mem_validate` functions, is used to ensure that a specific page is
readable before actually accessing a specific address from that page (in
order to avoid generating a segmentation fault). It does so by writing
the first byte of the page to a pipe. If the page is mapped readable,
the write succeeds, while if the page is not mapped, or not writeable,
then the write will fail with EFAULT or similar.

It seems that there has previously already been issues with ASAN here -
rather than using the `write` function in libc,  `syscall` is used in
an attempt to circumvent the memory checks from ASAN.

This patch changes the `write_validate` function to take the original
pointer, rather than the page-aligned version, and writes from that
specific address. Given that we are about to read that value anyway,
(probably to retrieve the contents of the stack in order to fulfill some
location expression in the dwarf data), the data at that location should
actually be initialized, while the data at the page-aligned offset may
not be.  This removes UMR errors reported by valgrind, and possibly other
tools that understand the semantics of individual numbered system calls
invoked via syscall (We can probably actually remove the syscall() variant
of the code, but I've left as-is to reduce the amount of churn for now)

While here, I've also fixed an issue where only the first page containing
the data accessed by access_mem is validated as part of `write_validate` -
`validate_mem`/`write_validate` now take a size argument, and we instead
walk through the content of the value to be verified, validating acces
every page-size forward of the first address, or, if that exceeds the
extent of the object, the last byte in the object. This function is
only used with sizeof (unw_word_t) as a size, so the pointers can only
possibly straddle a single page boundary, but it's probably clearer to
write it for the general case.  (See issue #232)

This addresses the issue for x86_64. Once done, I can do similar for
other architectures using the `write_validate()` mechanism